% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.types.R
\name{agg.ADAGES}
\alias{agg.ADAGES}
\title{Aggregation by ADAGES}
\usage{
agg.ADAGES(Shat.list, p)
}
\arguments{
\item{Shat.list}{list of K elements containing the selection sets.}

\item{p}{number of original variables of the model.}
}
\value{
A list containing following components:
\item{Shat}{aggregated selection set.}
\item{c}{optimal threshold.}
\item{K}{number of aggregated sets.}
}
\description{
This function aggregates a list of selection sets by ADAGES.
}
\details{
The function can be used in combination with \code{\link{multi.knockoffs}} and
\code{\link{multi.knockfilter}}.

The function does not have to be used in the context of multiple knockoffs.
It can also be used to aggregate the selection sets of any FDR controlling procedure
in general.

Applies the minimization of the complexity ratio as a criterion to determine the optimal threshold.
}
\examples{
#Multiple knockoff example
n <- 400; p <- 200; s_0 <- 30
amplitude <- 1; mu <- rep(0,p); rho <- 0.25
Sigma <- toeplitz(rho^(0:(p-1)))
X <- MASS::mvrnorm(n, mu, Sigma)
nonzero <- sample(p, s_0)
beta <- amplitude * (1:p \%in\% nonzero)
y <- X \%*\% beta + rnorm(n)

Xk <- multi.knockoffs(X, K = 5)
multi.res <- multi.knockfilter(X, Xk, y)
agg.ADAGES(multi.res$Shat.list, p = p)


#General example (selection sets with indices between 1 and 30)
Shat.list <- list(s1 = c(2,4,3,1,20,30), s2 = c(3,30,23,1,4,8),
                 s3 = c(3,4,5,13,15,12, 23:29, 30), s4 = c(1:10, 13:15, 17),
                 s5 = c(15:20, 23))
agg.ADAGES(Shat.list, p = 30)


}
\references{
Gui (2020). \emph{ADAGES: adaptive aggregation with stability for distributed feature selection}.
Proceedings of the 2020 ACM-IMS on Foundations of Data Science Conference.
\url{https://arxiv.org/pdf/2007.10776.pdf}
}
