% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.types.R
\name{agg.union}
\alias{agg.union}
\title{Union of selection sets}
\usage{
agg.union(Shat.list)
}
\arguments{
\item{Shat.list}{list of K elements containing the selection sets.}
}
\value{
Aggregated union set of selected variables.
}
\description{
This function aggregates a list of selection sets by their union.
}
\details{
The function can be used in combination with \code{\link{mult.knockoffs}} and
\code{\link{mult.knockfilter}}.

The function does not have to be used in the context of multiple knockoffs.
It can also be used to aggregate the selection sets of any FDR controlling procedure
in general.
}
\examples{
#Multiple knockoff example (union knockoffs)
n <- 400; p <- 200; s_0 <- 30
amplitude <- 1; mu <- rep(0,p); rho <- 0.25
Sigma <- toeplitz(rho^(0:(p-1)))
X <- MASS::mvrnorm(n, mu, Sigma)
nonzero <- sample(p, s_0)
beta <- amplitude * (1:p \%in\% nonzero)
y <- X \%*\% beta + rnorm(n)

Xk <- mult.knockoffs(X, K = 5)
q <- 0.2/2^((1:K)-1)
mult.res <- mult.knockfilter(X, Xk, y, q = q)
agg.union(mult.res$Shat.list)


#General example (selection sets with indices between 1 and 30)
Shat.list <- list(s1 = c(2,4,3,1,20,30), s2 = c(3,30,23,1,4,8),
                 s3=c(3,4,5,13,15,12, 23:29, 30), s4 = c(1:10, 13:15, 17),
                 s5 = c(15:20, 23))
agg.union(Shat.list)


}
