% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiknockfilter.R
\name{mult.knockoffs}
\alias{mult.knockoffs}
\title{Construction multiple knockoff matrices}
\usage{
mult.knockoffs(X, K, knockoffs = create.second_order)
}
\arguments{
\item{X}{n x p matrix of original variables.}

\item{K}{number of knockoff matrices.}

\item{knockoffs}{function for the knockoff construction. It must take the n x p matrix as input
and it must return a n x p knockoff matrix. Either choose a knockoff sampler of
the \code{knockoff} package or define it manually. Default: \code{create.second_order} (see below).}
}
\value{
A list that contains the K knockoff matrices.
}
\description{
This function generates a specified number of knockoff matrices given the data
and a knockoff sampling function.
}
\details{
This function requires the installation of the \code{knockoff} package prior to its execution.

The default knockoff sampler \code{create.second_order} is the second-order Gaussian knockoff construction from
the \code{knockoff} package.
}
\examples{
n <- 400; p <- 200
mu <- rep(0,p); rho <- 0.25
Sigma <- toeplitz(rho^(0:(p-1)))
X <- MASS::mvrnorm(n, mu, Sigma)

# Basic usage with default arguments
Xk <- mult.knockoffs(X, K = 5)
Xk

# Advanced usage with customized knockoff construction (equi-correlated)
equi.knock <- function(X) create.second_order(X, method = "equi")
Xk <- mult.knockoffs(X, K = 5, knockoffs = equi.knock)


}
\references{
Candes, Fan, Janson, and Lv (2018). \emph{Panning for gold. model-X knockoffs for high
dimensional controlled variable selection}. Journal of the Royal Statistical Society:
Series B (Statistical Methodology) 80(3), 551-577.
}
