% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.types.R
\name{agg.pKO}
\alias{agg.pKO}
\title{Aggregation by p-value knockoffs}
\usage{
agg.pKO(W.list, q = 0.2, gamma = 0.3, offset = 1, method = "BH", pvals = F)
}
\arguments{
\item{W.list}{n x p matrix of original variables.}

\item{q}{nominal level for the FDR control. Default: 0.2.}

\item{offset}{either 0 (knockoff) or 1 (knockoff+). Default: 1.}

\item{method}{the FDR controlling method in the last step. Either \code{"BH"} (default) or \code{"BY"}.}

\item{pvals}{if the aggregated p-values should be reported (logical). Default: \code{FALSE}.}
}
\value{
A list containing following components:
\item{Shat}{aggregated selection set.}
\item{B}{number of knockoff matrices.}
\item{pvals}{if specified, vector of aggregated p-values.}
}
\description{
This function only runs the aggregation. Given a list containing B elements
with the score vectors of length p, it computes aggregated p-values and applies
either BH or BY in the last step.
}
\details{
This function should be used in combination with \code{\link{mult.knockoffs}} and
\code{\link{mult.knockfilter}} (see example).
}
\examples{
n = 400; p = 200; s_0 = 30
amplitude = 1; mu = rep(0,p); rho = 0.25
Sigma = toeplitz(rho^(0:(p-1)))

X = MASS::mvrnorm(n, mu, Sigma)
nonzero = sample(p, s_0)
beta = amplitude * (1:p \%in\% nonzero)
y = X \%*\% beta + rnorm(n)

# Construction of K knockoff matrices
Xk <- mult.knockoffs(X, K = 20)

#Multiple knockoff filter
mult.res <- mult.knockfilter(X, Xk, y)

pKO.res <- agg.pKO(mult.res$W.list)
pKO.res

}
\references{
Benjamini and Hochberg (1995). \emph{Controlling the False Discovery Rate: A Practical
and Powerful Approach to Multiple Testing}. Journal of the Royal Statistical Society. Series B (Methodological) 57(1), 289-300.

Benjamini and Yekutieli (2001). \emph{The control of the false discovery rate in multiple testing under dependency}.
The Annals of Statistics 29(4), 1165-1188.

Meinshausen, Meier and Buehlmann (2009). \emph{p-Values for High-Dimensional Regression}.
Journal of the American Statistical Association 104(488), 1671-1681.

Nguyen, Chevalier, Thirion and Arlot (2020). \emph{Aggregation of Multiple Knockoffs}.
Proceedings of the 37th International Conference on Machine Learning.
\url{https://arxiv.org/abs/2002.09269}
}
